Milestone: No custom init code neccessary!
